+++ app/src/main/java/com/rickyhu/hushkeyboard/ui/keyboard/buttons/NotationKeyButton.kt
@@ -1,4 +1,4 @@
-package com.rickyhu.hushkeyboard.ui.keyboard
+package com.rickyhu.hushkeyboard.ui.keyboard.buttons
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.Orientation
